requirements/base.txt
django==1.8.15
https://github.com/hkmshb/django-ezaddress/archive/v0.3.1.tar.gz
django-filter==0.15.2
djangorestframework==3.3.3
mongoengine==0.10.6
django-rest-framework-mongoengine==3.3.0
